Move aggregators to sql/api
